Fixed divceil to prevent lose of accuracy

+++ include/nt2/toolbox/arithmetic/function/scalar/divceil.hpp
@@ -39,8 +39,7 @@ namespace nt2 { namespace ext
     NT2_FUNCTOR_CALL(2)
     {
       typedef typename NT2_RETURN_TYPE(2)::type type;
-      return iceil(float(a0)/float(a1));
-
+      return iceil(double(a0)/double(a1));
     }
   };
 } }
